backend/requirements.backend.txt
httpx==0.28.1
fastapi==0.121.2
uvicorn==0.38.0
SQLAlchemy==2.0.44
aiosqlite==0.21.0
redis==7.0.1
chromadb==1.3.4
numpy==2.3.5
PyPDF2==3.0.1
python-docx==1.2.0
langchain-text-splitters==1.0.0
sentence-transformers==5.1.2
transformers==4.57.1
tokenizers==0.22.1
torch==2.9.1
httpx==0.28.1
tqdm==4.67.1
pydantic==2.12.4
python-multipart==0.0.20
python-dotenv==1.2.1
orjson==3.11.4